slotBooker/pyproject.toml
[tool.poetry]
name = "slotBooker"
version = "0.1.0"
description = ""
authors = ["Sebastian Blum <sebastian.blum@alexanderthamm.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.3.0"
DateTime = "^4.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
slotBooker = "slotbooker.booker:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
